back/requirements.txt
Django
djoser
drf-yasg
django-filter
django-jazzmin
djangorestframework
django-cors-headers
psycopg2-binary
uvicorn